pyproject.toml
[build-system]
requires = ["setuptools>=75.5.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quik-tracy"
version = "0.3.0"
description = "A comprehensive tool for Tracy profiling with Docker and local support - build, capture, analyze, and export Tracy profiling data"
dependencies = ["rich_click", "docker", "pandas", "tables"]

[project.scripts]
quik-tracy = "quik_tracy.__main__:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 140
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
known_first_party = [
    "audiorenderer",
    "fonthelper",
    "mediaresolver",
    "stupeflix_ffmpeg",
    "quikavengine",
    "quiktasks",
    "quiktaskssystem"
]
profile = "black"
force_sort_within_sections = true
line_length = 140